set `TimeEvolution` Module as packag extention for OrdinaryDiffEq
From now on  the Module is only loaded when `OrdinaryDiffEq` is also loaded

The Package `PackageExtensionCompat` is used to make sure that HB stays backwards compatible by using `Requires`

Resolves #100, #98
