+++ src/HarmonicBalance.jl
@@ -12,32 +12,6 @@ module HarmonicBalance
     import Distances
     # using SnoopPrecompile
 
-    import Base: show, display; export show
-    export *
-    export @variables
-    export d
-
-
-    import Base: ComplexF64, Float64; export ComplexF64, Float64
-    ComplexF64(x::Complex{Num}) = ComplexF64(Float64(x.re) + im*Float64(x.im))
-    Float64(x::Complex{Num}) = Float64(ComplexF64(x))
-    Float64(x::Num) = Float64(x.val)
-
-   # default global settings
-   export IM_TOL
-   IM_TOL::Float64 = 1E-6
-   function set_imaginary_tolerance(x::Float64)
-       @eval(IM_TOL::Float64 = $x)
-   end
-
-   export is_real
-    is_real(x) = abs(imag(x)) / abs(real(x)) < IM_TOL::Float64 || abs(x) < 1e-70
-    is_real(x::Array) = is_real.(x)
-
-    # Symbolics does not natively support complex exponentials of variables
-    import Base: exp
-    exp(x::Complex{Num}) = x.re.val == 0 ? exp(im*x.im.val) : exp(x.re.val + im*x.im.val)
-
     include("types.jl")
 
     include("utils.jl")
@@ -52,17 +26,12 @@ module HarmonicBalance
     include("saving.jl")
     include("transform_solutions.jl")
     include("plotting_Plots.jl")
-    include("hysteresis_sweep.jl")
 
     include("modules/HC_wrapper.jl")
     using .HC_wrapper
 
     include("modules/LinearResponse.jl")
     using .LinearResponse
-
-    include("modules/TimeEvolution.jl")
-    using .TimeEvolution
-    export ParameterSweep, ODEProblem, solve
 
     include("modules/LimitCycles.jl")
     using .LimitCycles
@@ -71,6 +40,12 @@ module HarmonicBalance
     using .KrylovBogoliubov
     export first_order_transform!, is_rearranged_standard, rearrange_standard!, get_equations
     export get_krylov_equations
+
+    using PackageExtensionCompat
+    function __init__()
+        @require_extensions
+    end
+    export ParameterSweep, ODEProblem, solve, follow_branch
 
     # precomp_path = (@__DIR__) * "/../test/"
     # @precompile_all_calls include(precomp_path * "parametron.jl")
